+++ src/RestEaseUnitTests/ImplementationBuilderTests/RequesterPropertyTests.cs
@@ -1,0 +1,29 @@
+﻿using Moq;
+using RestEase;
+using RestEase.Implementation;
+using System;
+using System.Collections.Generic;
+using System.Net.Http;
+using System.Net.Http.Headers;
+using System.Text;
+using System.Threading.Tasks;
+using Xunit;
+
+namespace RestEaseUnitTests.ImplementationBuilderTests
+{
+    public class RequesterPropertyTests
+    {
+        public interface IHasRequesterProperty
+        {
+            IRequester Requester { get; }
+        }
+
+        [Fact]
+        public void HandlesRequesterProperty()
+        {
+            var requester = new Mock<IRequester>().Object;
+            var client = RestClient.For<IHasRequesterProperty>(requester);
+            Assert.Equal(requester, client.Requester);
+        }
+    }
+}